Add files via upload
Problem: Significant problem that can arise  when creating a health webpage is ensuring the accuracy and reliability ot the information provided.  Health information must be evidence-based and presented by credible sources to build trust with the users.

Solution: Maintaining the accuracy and reliability of health information is essential to ensure the webpage servers as a trustworthy resources for users seeking medical advice and information.